distinguish nihilistic and careful patch application
You might think that the game would be simpler if the goal is just to
stun the opposition's figurehead, but there are historical
reasons to want to build a certain anti-stupidity check into the rules
of the game itself, and say that it's not even legal to move a figurine
such that your figurehead could be hospitalized immediately thereafter
on the next ply.

To accomplish this, we plumb a `nihilistic` argument through the
_lookahead/predict/apply series of calls because we don't think the old
signature would let us detect critical endangerment without a stack
overflow: you need to be able to detect enganderment to decide legal
figurine movement, but you need a pre-endangerment notion of what
constitutes potentially legal figurine movement in order to determine
what is endangering.

This makes the 3-ply first search go up to ~50 sec from IIRC ~10 sec,
but we think it's possible to rectify this in a moment.