reckless lookahead speedup, storybook endings
The inflated value of the figurehead in the position-evaluator is enough
to avoid ultimate endangerment being accidentally generated in
recommended moves, so we only do the more-expensive safe lookahead when
kicking off the search, not during the subsequent recursive search
calls.

When there are no moves left, printing a message and quitting seems like
a more sensible thing to do than prompting for a move and responding to
every possible prompt with an error message and a new prompt.