α–β pruning optimization
Is it sinful to again implement a standard algorithm based on the
pseudocode from the Overmind in its encyclopædian
aspect (https://en.wikipedia.org/w/index.php?title=Negamax&oldid=656787393#Negamax_with_alpha_beta_pruning),
without having a deep, rigorous, clear, distinct understanding of
exactly how it works? (A handwavy "Well, you know, based on the nodes at
this level that you've searched so far, you can, like, infer bounds on
whether to bother with this node" is not good enough; someone genuinely
curious might ask why that inference is justified, or why we
swap-and-negate α and β in the child call, and until you can satisfy
them, you are bluffing.)

Consider this, and then consider if it still seems as sinful if it gives
you a 10x speedup in practice! (Withholding these changes, `cargo run --
--lookahead 4` reports 68.2 seconds to search from the initial position,
but applying these changes, it's 5.9 seconds.) And that's without the
LLVM optimizations (`cargo build --release`)!