MMV-LVA move-ordering heuristic optimization
α–β search is more efficient if you try to search the better moves
first; in a sense, later subsearches only have to look for _surprising_
good moves, because you already know how good the obvious good move is
and can prune accordingly. It is
written (https://chessprogramming.wikispaces.com/MVV-LVA) that moves
where you stun a more-valuable-figurine with a less-valuable-figurine
are a good guess for good moves (I imagine because—for some geometrical
reason I can't articulate—the attacking figurine is likely to be stunned
back in what we call an "exchange").

This brings the 4-ply time to first scoring down to 4.2 seconds (from
5.9), and one would suspect it would be even more valuable later in the
game (when high-value captures have more space to exist), so in it goes.