transposition memoization optimization
It's possible to reach the same game state by different move sequences:
when figurines move on different sides of the board and don't interact,
we don't care what order the moves happened in. If we memoize the values
of positions we've seen before, we don't have to search them
again. (This requires what seem like gymnastics to a Pythonista in order
to let the borrow checker bless our use of the shared hashmap, but it's
not so bad.)

You might be worried about how this interacts with negamax (where we're
continually negating values in order to not have to keep track of the
"minimizing" and "maximizing" team separately), but since `to_move` is
part of the WorldState memoization keys, it should be fine.

A quick manual benchmarking consisted of selecting the first ("0.") move
for each of the six plies both with and without this patch, and jotting
down the reported thinking times. The observed timings were
around (4.21, 4.91, 3.82, 7.58, 4.46, 9.44) and (2.94, 3.61, 2.67, 5.54,
3.14, 6.87), for an average improvement of almost 40%!!