core towards correspondence
Console games are fun, but have a limited audience and somewhat
constrained user-interfaces. I'm imagining it would be instructive to
fashion the Leafline binary as the stateless core of a webapp wrapped
around it: POST a serialization of the current world-state, receive a
response of the world-state after the AI's reply, which will be rendered
using JavaScripts. Now that we have serialization (preserve/reconstruct)
and this command-line flag, all that would be left to do is build the
webapp.