new design Foundations; or, Bootstrap is so overrated
Seriously. It's not that Bootstrap is bad (well, except they do violence
to our hopes for meaninfully semantic HTML with their bizarre
overloading of venerable tags like <i>; so, yes, Bootstrap is bad), so
much as that, for this project, maybe we don't want to use the same
frontend framework as every newbie web developer and their dog, who slap
Twitter Bootstrap on their nugatorious cruddy CRUD apps to make it
momentarily less obvious that they don't know design.

Leafline is different. Leafline is cutting-edge. Leafline's build
scripts are in Python 3; its frontend logic is in ECMAScript 6 (!), its
backend is a Clojure (!!) app that wraps a Rust (!!!) binary. Do you
_really_ think we're going to use Bootstrap for our CSS?