small name/commentary/code tweaks
Copy-pasting long comments verbatim is sloppy; if something needs to
be noted again, be brief in the secondary location.

We can use the #[ignore] attribute in preference to commenting out a
benchmark. (Also, wow, we have block comments? The book doesn't seem
to mention them, so I assumed they didn't exist!)

(I had thought the recommended name for the unit test modules was
"test", but it looks like "tests" is now the standard as of April's
rust-lang/rust@07cc7d99!)