restoration and preservation of secret service eligibility
Note that web client play still doesn't understand secret service. You
would hope that you could perform this special move by dragging the
figurehead two squares. But as things stand, our client is too dumb for
this to work; it is no doubt something to think about when we get to the
point of actually validating figurine movements in web client play.