defer to few remaining warnings; run out of digits
Imposing snake_case on crate names seems excessive to me—the name of a
program is a proper noun that should be capitalized, and isn't it
desirable for the technical crate name to be the same as the name of the
program? But I couldn't bear to slap #![allow(non_snake_case)] on the
entire crate, which I'm imagining would be the other remedy. Also, I
think the macro_export attribute is for exporting to other crates, which
doesn't concern us?—because Leafline, as implemented in the now
lowercase-l "leafline" crate, is an application rather than a library.