dubious format pass
It's not on crates.io, but nrc/rustfmt is already a usable
code-formatter. The changes in this commit come from running this tool
over the codebase and then doing a manual pass of fixing things that
bothered me.

With Rust, I'm in the unfortunate situation that how I want my code to
look now, what I might want my code to look like sometime later after I
get around to investing in a bigger monitor or tweaking my .emacs (at
home, my Emacs windows are 80 characters wide), what rust-mode (tabbing
or C-M-\) wants my code to look like, and what this new tool wants my
code to look like, are all subtly or not-so-subtly different. The
ultimate resolution of this tension is yet unknown.