Minor cleanups; adding a start state to cmdline invocation
I noticed while trying to test the smartness of the ascension prediction
that when a servant ascends (or is brevetted), the servant is still on
the board afterward!! this is a pretty bad logic bug but tbh it's
troubling enough that I don't have the energy to fix it right now.

The solution *might* be to do as we did for secret service manouevres,
and have the patch be an illegal move, something like "Princess from a7
to a8" (when of course there is actually a _servant_ on a7) and then
figuring it out magically in `apply`. but since `aspire_maybe` is
already attempting to do this in a different place, I'm not so sure.

Anyway, here's a test that will be fixed once servant ascension works
properly.