fix spontaneously spawning ascendant doppelgängers bug; fn rename
The present author must have been pretty sleepy last night to have typed
`self` when what was actually meant was `premonition.tree`! Special
thanks to Alexander Corwin for the bug report (and the test case
committed in e82e9f85).

Also, ending a function name with "_maybe" makes it sound like it
returns an Option; the ad hoc subroutine that handles appending
premonitions to the lookahead vector (and, notably, appending a separate
premonition for each possible outcome of an ascension) must be called
something else.