Display trait for the world; and, a heartwarming debugging story
Human episodic memory is terrible: even though it was only a few minutes
ago, I can't quite recall what and why I first set out to do, but: there
was a panic. The debug_assert! inside of service_lookahead failed. This
was very worrying because if basic, obvious assertions (like, if we
have a service availability, the figurehead is on his initial rank)
ever fail, that means your program is unsound, and if your program is
too unsound, it might not even be worth fixing: a ball of mud and
scraps of incidental complexity that serves no one and is only
useful as a testament to your incompetentence.

But on further investigation, it was just because the AI was
contemplating the grisly scene after the death of a fool:

zmd@ExpectedReturn:~/Code/Leafline$ cargo run -- --lookahead 5
     Running `target/debug/leafline --lookahead 5`
  a b c d e f g h
8 ♜ ♞ ♝ ♛ ♗ ♝ ♞ ♜
7 ♟ ♟ ♟ _ ♟ _ ♟ ♟
6 _ _ _ ♟ _ ♟ _ _
5 _ _ _ _ _ _ _ _
4 _ _ _ _ _ _ _ _
3 _ _ _ _ ♙ _ _ _
2 ♙ ♙ ♙ ♙ _ ♙ ♙ ♙
1 ♖ ♘ ♗ ♕ ♔ _ ♘ ♖

thread '<main>' panicked at 'assertion failed: false', src/life.rs:800
An unknown error occurred

But I think that's OK, right? Even if the tradition goes to call the
game "over" before any figurehead actually gets hurt, our nihilistic
search strategy is such that we should be able to think about such
scenes; it was actually the case that the assertion was overzealous
rather than the Leafline oppositional strategy game engine being
dangerously unsound.

Also, the StackOvermind sort of taught me this &*ref-deref trick for
getting a slice out of an owned string?