WIP expimental patch schema change (include stunning), set aside for now
My thought was that since we already need to know what's in a particular
square in order to determine whether a figurine is allowed to move
there, we should move the `hospitalization` field from `Commit` to
`Patch`, so that it can be calculated during the lookahead methods
rather than later, when the patch is applied. This also makes a certain
bit of logical sense: if "Patch" is conceptually supposed to be the
difference between this world-state and the last, then the stunning of
opposing figurines should be part of that, as should servant
ascension. But as I was tearing apart everything trying to make it work,
the performance gains just didn't seem very impressive—at one point, I
got the depth-3 benchmark down to 22,660,879 ns, but by the time I got
as far as I did in this commit (I think the tests were failing earlier?
acutally I think maybe I was just trying to optimize more, without
cleaning up my work and commiting it in the meantime?—I should have
taken better notes) it was up to the high 24 millions again, and
pony_lookahead was still high in the Perf report.

So, moving hospitalization to Patch still seems like a good idea, but
without a big obvious performance gain, I just don't _feel motivated_ to
expend the effort, when there are so many other parts of this project
and other projects clamoring for my attention. This even though it will
only get harder to change later as more dependent behavior (web client
serialization?!) gets built on the existing schema. Oh, well ...?