liven identifiers; prune old function; don't score spuriously precisely
If at some point we are interested in the original negamax
search (perhaps to be able to say quantitatively how optimized the
Leafline AI has gotten compared to its beginnings), we'll dig it up from
the Git history then.

Is there a way to say "display at most this many decimal places, but
omit them if you would have printed '0'?" But just printing the .0 is
better than having to see scores like "-3.8007813" due to floating-point
roundoff.