super naïve iterative deepening
I hate the code in this commit, but it's what I came up with towards the
goal of giving the AI a bounded amount of thinking time rather than a
fixed number of levels (the consistency providing a more pleasant
user-experience than easy moves taking 2 seconds but hard ones taking an
unpredictable amount of time that occasionally breaks a minute).

Iterative deepening is said to be not as bad as it looks, even when done
as naïvely as here: the nature of exponential growth is such that the
last level searched is always the most expensive by far. The
sophisticated way, however, is said to be actually really good because
you can exploit what you learned from earlier levels to find α–β cutoffs
really quickly. Maybe we'll experience this ourselves someday.

I still love Option types, but I am beginning to question the ergonomics
of using them where in Python I would use a keyword argument that
defaults to None.

Putting the time cutoff in potentially_timebound_kickoff rather than
deeper in the stack of search functions doesn't result in as quick of a
cutoff as you might hope (scoring alternatives took 11529 ms), but I
didn't want to tear the guts apart so soon and on a Sunday.

I don't understand the motivation for allowing tuple-destructuring
assignment with `let` but not without.