client requests bounds on thinking time and depth
The checking for the deadline in the premonitions loop of
potentially_timebound_kickoff doesn't look as good-enough as I
thought—it's usually not more than a couple seconds late, but I've seen
it go way over a few times (no doubt sensitive to the time request and
the position, sneaking under the deadline before searching a premonition
to depth 6 being much worse than 5). I suspect the real answer involves
running the minimax search in a separate thead (with the right channels or
whatever so that it can report its intermediate results), not entirely
unlike the real way to do iterative deepening is to take advantage of
the low-depth searches to do move ordering. Might we imagine that my
ignorance on some September Sunday evening might be forgiven?