nix ultimate-endangerment short-circuit until it is better understood
Alexander Corwin reported what appears to be a bug in the play engine:
from `--depth 5 --from
"r1bq1rk1/ppppnppp/5n2/2b1p3/2PP4/2N1P1P1/PP2NPBP/R1BQK2R b -"` (a
position encountered during a game between Clay-in-the-Shadow-of-Orion
and Young Money), Leafline reported the top movement as `0. ♝ from c5
to d4, stunning ♙ (score 20004.2)`, suggesting that it foresees a
critical endangerment winning the game for the Blue Team. But after
the exchange takes place, the apparent victory opportunity vanishes
(`0. ♞ from e7 to c6 (score -2.7)`)! Without the short-circuiting
introduced in c88b40aa, the recommended movement is `0. ♟ from e5 to
d4, stunning ♙ (score 2.1)`, which seems more reasonable.

It is sad and peculiar that the authors' understanding of α–β search
could be so mistaken that an optimization and rationale that seemed
correct (if α is in the vicinity of the figurehead's worth, a sure
victory is within the planning horizon, so we can stop searching)
could turn out to be wrong, but until we have a better understanding
of exactly if and why it was wrong and how to accomplish the same
purpose correctly, let's stop trying to be clever and remove the
short-circuiting for now.

(The present author is actually kind of tempted to say that the
naïve behavior of always seeking material advantage without regard to
the traditional victory condition has the potential to be fun and
dramatic, if it results in the AI "running up the score" stunning lots
of figurines in preference to cutting to the obvious victory, but
that's a separate matter that doesn't sound like it would happen
automatically.)