searching in a thread—but at what cost, for what gain?
This commit demonstrates that we can call α_β_negamax_search in a
spawned thread. Not that it does us any good here, because the borrow
checker won't let us lend it the hashtable. But perhaps this is a
starting point from which we can do something interesting and useful.

We also slap Copy back onto Commit and Patch, but I think that's OK
maybe (I thought I remembered reading somewhere that even when your type
has move rather than "copy" semantics, the metal is still necessarily
_copying_ memory; it's just an abstraction to prevent you from copying
pointers that can be used to mutate their reference; if this
understanding is correct, Commit and Patch should have the Copy-nature
because their components are—in the present author's opinion— value-like
rather than identifiably-persistent-object-like.)