poll channel to sharpen iterative deepening timeout; depth display fix
I was imagining that there would be some non-blocking analogue of
std::thread::JoinHandle.join(), but the docs didn't seem to mention one.

The depth variable in `iterative_deepening_kickoff` gets incremented to
prepare for the next search layer (the last of which will almost
certainly not get a chance to finish), so we actually have to decrement
what we return in order to not mislead.