use more cores
We had been searching each of the kickoff-level premonitions in a thread
for timeout purposes (because I really really wanted the AI movements to
take a consistent, predictable amount of time), but then synchronously
waiting for each of them. We could imagine someone arguing that this is
entirely missing the point. If instead we can kick off all the threads
and then poll their channels in a loop, we use both cores of my PC, and
it looks like we get a performance boost—although perhaps not quite as
dramatic and easy to measure as some might have hoped for. The first
scoring in the 6-ply terminal game goes down to about 5.7 seconds, from
8.9 on the parent commit. The depth-three kickoff benchmark goes down to
23,005,870 ns/iter (+/- 3,487,832), from 47,771,309 ns/iter (+/-
7,058,747) on the parent commit. On master, the 6-ply terminal scoring
was about 8.5 seconds—so you would _think_ that the 30+% improvement
really is a net improvement, not just an amelioration of earlier
development on this branch having broken performance—but the depth-three
benchmark was 25,562,534 ns/iter (+/- 169,301) (still worse than what we
have here, but only usually, and not by much).

So, I'm not entirely sure what's going on here and exactly how the
performance characteristics of the new code contrasts under different
workloads, but I like it overall. And we can hope that it'll be even
better on a machine with more cores than the circa 2008 hardware that
this commit was authored on.

I learned about the "iterate over indices backwards in order to mutate
the iterable with swap_remove" trick from Alexis Beingessner's talk at
RustCamp 2015 (briefly recounted in the eleventh paragraph of
http://zackmdavis.net/blog/2015/08/rustcamp-reminiscences/; slides at
http://cglab.ca/~abeinges/talks/iter/).