cops like to advance to the seventh, servants aspire to ascend
I heard that the analogues of orange (respectively blue) cop-like
figurines in other, similar oppositional strategy games like to live on
what we would call the seventh (respectively second) rank. And I suspect
that the 4- to 6-ply lookahead that we can get with reasonable waiting
time isn't enough to properly incentivize servant ascension, so we add a
bonus for servants on the sixth (respectively third &c.) and seventh
ranks. But these specific figures are just guesses, possibly too
conservative guesses.