a world that runs on JavaScript is a world headed for destruction
I write Python in order for the world economy to consider me worthy of
life, and hadn't permitted myself to imagine a bug like this (property
access of "triumpant" with no 'h'). Surely that's an AttributeError! Or
a KeyError! It can't just be undefined and fall through to the 'else'
clause! But it can. It will. Then Therac-25 administers a lethal dose of
radiation, and planes fall from the sky.