recognize option value of secret service eligibility
As recounted in the commit message of swiftstack/Chesswork@c268d3694, a
recent 7-ply search included moving the figurehead among its top-ranked
movements. But even a human can see that the ability to call in the
secret service sometime in the future is useful, even if we don't
foresee needing it in the next 5, 6, or 7 plies. Let's add a (very
small—only one deciservant) bonus to ensure that we don't throw it away
without some kind of reason.