inventory available memory (no servitude to Cupertino or Redmond)
Deep searches have been observed to take up enormous amounts of
memory. I don't currently know how much of it is inevitable stack space
and how much of it is our transposition table, but it seems plausible
that we can and should be kinder to our host by limiting the size of the
table (so we don't fill up memory and get killed or start swapping),
without sacrificing our ability to gain from looking at the table, and
that this would be assisted by glancing at how much memory our
substrate has to offer us.