if we can't do it in 1.5 GiB, maybe it's not worth doing
A quick experiment shows that bounding the size of the déjà vu table
trades memory for time in the way you would expect: a 7-ply search from
the initial condition with this 1.5 GiB hardcoded limit took 108.7
seconds and I think I remember seeing its `top`-reported memory
percentage creep above 30, whereas when it was bumped down to 0.2, it
took 220.7 seconds, but I don't think the memory usage went above 5%.

What is on the other hand not expected, is that the scored movement
lists of those two runs are noticeably different—but the déjà vu cache
is only supposed to save us time, not affect the results! What could we
be doing wrong?!

The hardcoded 1.5 is an arbitrary placeholder, we probably want to
govern this with a CLI flag (perhaps with an option to use a the
"percentage of the detected free memory" strategy?).