rework search return values, display principal variations
We don't need to return a commit from the search routine: the caller
already knows which premonition's world state was passed in to score,
and doesn't need (and wasn't using) the full state of the grandchild
node. But we would like to keep track of the patch chain to the
grandchild and further down the tree.