require déjà vu cache hits to be at the same search depth
So, the engine has been observed to do seemingly pretty stupid things
very occasionally, and not even issue scores deterministically every
time, and it's very strange, because it plays well enough the vast
majority of the time such that you would think the authors got α–β
pruning correct, so what could be wrong? Without understanding the exact
bug in detail, the present author has reasons (reasons that may not be
worth the writing effort to explain in this commit message) to believe
this might help.