require déjà vu cache hits to have equal or greater subtree soundness
The thought is that if the same world-state appears in the tree at
different depths (which would have to be due to a figurine's movement
one way and then back again, but whatever), they'll in general be
assigned different scores, and the deeper appearance will have a
less-accurate score derived from a shallower subtree beneath it between
it and the maximum search depth. This is a more precise idea than the
earlier vague feeling (98b969fb) amounting to "eh, let's require the
same search depth," and precise thought is important even when it
doesn't immediately make for a noticeable performance difference.