sometimes, with hashing speed, it's the thought that counts
It's kind of mostly lost in the noise of insignificance in our
benchmarks, but according to the table in the README of this new crate,
our WorldState should be big enough for this hash function to be
substantially better than the standard SipHash. Some might say that's
not our bottleneck and we don't care. But Leafline is cutting-edge;
we're not afraid to make an arbitrary technical decision in order to
impress the townies who live in Python (where, as a child, the grown-ups
almost never spoke of hash functions, and certainly never picked out a
custom one for our dictionaries!).