correct for initiative; plus, worries about losing grip on the codebase
We know there's an even-odd score oscillation that has to be in the map,
not the territory, and if we want to do non-even-parity quiessence
search extensions, you'd guess that we'd want scores at different depth
parities to be comparable. So, this _probably_ doesn't make things
worse?

Although ...

The integer overflow called out in a comment was pretty bad, and could
be related to a very recent sighting of the
thought-to-have-been-banished nondeterministic stupidity—and speaking of
nondeterminism, the super-impressive performance gain that I observed
just before committing the linktime optimization configuration didn't
seem to reproduce on a clean build! What's going on?!