establish a Python embassy
We now build libleafline.so, a shared library that exposes a `score`
function, as used in the new `embassy` Python module/script.

For "convenience."

This really seems to hurt our compile times, because it looks like
`cargo build` wants to compile both the .so and the excutable, even with
the `--bin leafline` or `--lib` options that one somehow got the idea
that it would have respected for this.